Remove Navbar
Remove the Navbar because we don't have any links. Add a vscode settings json so that search excludes the next.js cache stuff. Start abstracting out BarChart component and fake data into its own charts folder. Clean up TOS.

Resolves 73.
