+++ web/src/components/charts/types.ts
@@ -1,0 +1,7 @@
+export interface IText {
+  key: number
+  name: string
+  text: string
+  length: number
+  words?: Array<string>
+}
